Added QuizController
