+++ app/src/main/java/com/example/felipe/geoquiz/Question.kt
@@ -1,5 +1,3 @@
 package com.example.felipe.geoquiz
 
-data class Question(var textResId: Int, var isTrue: Boolean) {
-
-}
+data class Question(var textResId: Int, var isTrue: Boolean)
