+++ app/src/main/java/com/example/felipe/geoquiz/QuizActivity.kt
@@ -8,49 +8,37 @@ import kotlinx.android.synthetic.main.activity_quiz.*
 
 
 class QuizActivity : AppCompatActivity() {
-    private val questionBank = arrayOf(
-            Question(R.string.question_australia, true),
-            Question(R.string.question_oceans, true),
-            Question(R.string.question_mideast, false),
-            Question(R.string.question_americas, false),
-            Question(R.string.question_africa, true),
-            Question(R.string.question_asia, true)
-    )
+    private val quizController = QuizController()
     private var index = 0
-
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_quiz)
 
-        txtQuestion.setText(questionBank.get(index).textResId)
+        txtQuestion.setText(quizController.getQuestion(index))
         btnNext.setOnClickListener {
-            index = (index + 1) % questionBank.size
-            txtQuestion.setText(questionBank.get(index).textResId)
+            index = (index + 1) % quizController.size
+            txtQuestion.setText(quizController.getQuestion(index))
         }
         btnPrev.setOnClickListener {
-            index = when ((index - 1) >= 0) {
-                true -> index
-                false -> questionBank.size - 1
+            index = when (index - 1 >= 0) {
+                true -> index - 1
+                false -> quizController.size - 1
             }
-            txtQuestion.setText(questionBank.get(index).textResId)
+            txtQuestion.setText(quizController.getQuestion(index))
         }
 
-        btnTrue.setOnClickListener { printResult(true) }
-        btnFalse.setOnClickListener { printResult(false) }
+        btnTrue.setOnClickListener { showAnswerResult(true) }
+        btnFalse.setOnClickListener { showAnswerResult(false) }
     }
 
-    private fun printResult(message: Boolean) {
-        val msg = when (message == questionBank.get(index).isTrue) {
+    private fun showAnswerResult(answer: Boolean) {
+        val answer = when (quizController.checkAnswerResult(answer, index)) {
             true -> R.string.correct
             false -> R.string.incorrect
         }
-        with(Toast.makeText(this, msg, Toast.LENGTH_SHORT)) {
+        with(Toast.makeText(this, answer, Toast.LENGTH_SHORT)) {
             setGravity(Gravity.TOP, 0, 200)
             show()
         }
     }
-
 }
-
-
-
